feat: mockery support, new UTs, Makefile, lint, fixing lint errors

+++ pkg/fetcher/store/types.go
@@ -43,7 +43,7 @@ func (ut *UnsyncedTopics) AddTopic(address common.Address, topic common.Hash, co
 func (ut *UnsyncedTopics) GetAddressesAndTopics() ([]common.Address, [][]common.Hash, uint64) {
 	addresses := make([]common.Address, 0, len(ut.addrToTopicCoverage))
 	topics := make([][]common.Hash, 0, len(ut.addrToTopicCoverage))
-	minCoveredBlock := uint64(^uint64(0)) // Max uint64
+	minCoveredBlock := ^uint64(0) // Max uint64
 
 	for addr, topicMap := range ut.addrToTopicCoverage {
 		addresses = append(addresses, addr)
